+++ Vector.cpp
@@ -1,27 +1,29 @@
 #include "Vector.h"
 
-Vector::Vector(const DefinedPoint point, const int initalPosition) : m_point(point), m_initalPosition(initalPosition)
+Vector::Vector(const DefinedPoint point, const double initalPosition)
+    : m_point(point),
+      m_initalPosition(initalPosition)
 {
 
     SetVectorCoordinate(m_initalPosition);
 }
 
-std::pair<int, int> Vector::CalculateVectorCoordinate(const int initalPosition) const
+std::pair<double, double> Vector::CalculateVectorCoordinate(const double initalPosition) const
 {
-    return std::pair<int, int>(m_point.getCoordX() - initalPosition, m_point.getCoordY() - ZERO);
+    return std::pair<double, double>(m_point.getCoordX() - initalPosition, m_point.getCoordY() - ZERO);
 }
 
-void Vector::SetVectorCoordinate(const int initalPosition)
+void Vector::SetVectorCoordinate(const double initalPosition)
 {
     m_vectorCoordinate = CalculateVectorCoordinate(initalPosition);
 }
 
-std::pair<int, int> Vector::GetVectorCoordinate() const
+std::pair<double, double> Vector::GetVectorCoordinate() const
 {
     return m_vectorCoordinate;
 }
 
-void Vector::Recalculate(int position)
+void Vector::Recalculate(double position)
 {
     SetVectorCoordinate(position);
 }
